update date selection options to checkout page
