+++ nougat/src/Admin/MainPage/MainContent/DeliverySettings/DateTime/TimeSettings/index.js
@@ -1,0 +1,11 @@
+import React from "react";
+
+import { Container } from "./styles";
+
+export default function TimeSettings() {
+  return (
+    <Container>
+      
+    </Container>
+  )
+}
