+++ nougat/src/Admin/MainPage/MainContent/DeliverySettings/DateTime/TimeSettings/styles.js
@@ -2,10 +2,6 @@ import styled from "styled-components";
 
 export const Container = styled.div`
   display: flex;
-  flex-direction: column;
-`;
-
-export const DateSection = styled.div`
-  display: flex;
   flex-direction: row;
 `;
+
