+++ nougat/src/redux/actions/dateTime.js
@@ -1,0 +1,9 @@
+export const getDate = (date) => ({
+  type: "GET_DATE",
+  payload: date,
+});
+
+export const getTime = (time) => ({
+  type: "GET_TIME",
+  payload: time,
+});
