+++ nougat/src/utils/noEnterSubmit.js
@@ -1,0 +1,5 @@
+export default function doNotSubmitOnEnter(event) {
+  if (event.key === "Enter") {
+    event.preventDefault();
+  }
+}
